#4 - Rename dirs + move public header files to include dir
