+++ include/Animation.h
@@ -4,9 +4,12 @@
 
 class Animation {
     public:
+        Animation();
+        ~Animation();
+
         ActorAnimator GetRoot();
         void LogRender(float framerate);
 
     private:
-        Actor mRoot;
+        Actor* mRoot;
 };
