+++ src/Actors/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_library(actors Actor.cxx BoxActor.cxx ActorData.h BoxData.h)
+add_library(actors Actor.cxx BoxActor.cxx)
 target_include_directories(actors PUBLIC "${PROJECT_SOURCE_DIR}" "${PROJECT_SOURCE_DIR}/../include")
