--- src/Animation/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(animation Animation.cxx ActorAnimator.cxx)
-target_include_directories(animation PUBLIC "${PROJECT_SOURCE_DIR}" "${PROJECT_SOURCE_DIR}/../include")
