--- src/Renderer/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(renderer LogRenderer.cxx)
-target_include_directories(renderer PUBLIC "${PROJECT_SOURCE_DIR}" "${PROJECT_SOURCE_DIR}/../include")
