+++ src/Renderers/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(renderers LogRenderer.cxx)
+target_include_directories(renderers PUBLIC "${PROJECT_SOURCE_DIR}" "${PROJECT_SOURCE_DIR}/../include")
