+++ src/Renderers/LogRenderer.cxx
