+++ src/Renderers/LogRenderer.h
