+++ src/main.cxx
@@ -1,4 +1,4 @@
-#include "Animation/Animation.h"
+#include "Animation.h"
 
 void CreateAnimation(ActorAnimator& root);
